+++ examples/3-team-guide/topic-team/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     axual = {
       source  = "Axual/axual"
-      version = "2.8.2"
+      version = "2.9.0"
     }
   }
 }